LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/unsupported-axis-expected.txt

Test

FAIL size(width > 0px) assert_equals: expected "true" but got ""
PASS size(height > 0px)
FAIL size((height > 0px) or (width > 0px)) assert_equals: expected "true" but got ""
FAIL size((width > 0px) or (height > 0px)) assert_equals: expected "true" but got ""
FAIL size((orientation: landscape) or (width > 0px)) assert_equals: expected "true" but got ""
FAIL size((width > 0px) or (orientation: landscape)) assert_equals: expected "true" but got ""
PASS size((height > 0px) or (orientation: landscape))
PASS size((height > 0px) or (orientation: landscape)), with contain:size
FAIL size(inline-size > 0px) assert_equals: expected "true" but got ""
PASS size(block-size > 0px)
PASS size(block-size > 0px), with writing-mode:vertical-rl
